App-Domains make your Apps popular
Apps make the mobile world go ‘round, and now they have their own more secure home on the web: the App-Domain, the most relevant domain extension for mobile apps.

You can use your App-Domain to showcase a unique and trustworthy destination, as a relevant download link, for deep linking, or for sharing screenshots, release notes, and reviews. Get your App-Domain now to start sharing your app with the world.

Key points about the TLD:
The App-Domain is an open TLD with a focus on improved security. The entire .app TLD has been added to the HSTS preload list, which means registrants will have to provision and set up SSL certificates in order for their content to be loaded in modern browsers

The App-Domain is a secure namespace, meaning that HTTPS is required for all .app websites. You can buy your App-Domain now, but in order for it to work properly in browsers you must first configure HTTPS serving.

The relationship between a better ranking in search engines and new top-level domains has already been proven in a study of Searchmetrics for .berlin domains. Websites with .berlin domains often rank better with regional searches in Google than sites with .de domains and .com domains. The result of the Search Metric study can be summarised as follows:
»In 42% of searches, .berlin domains rank better locally.«
Another study by Total Websites in Houston shows that the results of the Searchmetrics study are in principle applicable to all new top-level domains, including the new App-Domains. Total Websites notes that Google sees the domain extensions of new top-level domains as an important criterion for the evaluation of a domain and, therefore, concludes that:
»It's clear that the new top-level domains improve the ranking in search engines.«

Marc Mueller

http://www.domainregistry.de/app-domain.html (English)
http://www.domainregistry.de/app-domains.html (German)